requirements.txt
huggingface_hub==0.29.2
joblib==1.4.2
numpy==1.24.3
openai_whisper==20240930
protobuf==3.20.3
python-dotenv==1.1.1
streamlit==1.47.1
